Stop tracking players.xml
This file is in its final, empty form. Next developments are likely to fill this file with data either unsuitable for a serious use, or private data that shouldn't be shared online.
Users may replace this file by the one in their TOM data/ folder, as they will work the same.