Fix bug relative to event assigned to PlayersList
Due to a confusion on the meaning of "this", both lists were using the same event: clicking on a player in the registered list activated the event that tries to insert that same player in the registered list. As this had no influence on the software's result (a player already in that list cannot be readded), it was important to fix that, in order to not repeat the same scenario in a worse situation.